iOS: rename interactions to wait_for_condition
