+++ lib/calabash/ios.rb
@@ -17,7 +17,7 @@ module Calabash
     require 'calabash/ios/environment'
     require 'calabash/ios/application'
     require 'calabash/ios/device'
-    require 'calabash/ios/interactions'
+    require 'calabash/ios/wait_for_condition'
     require 'calabash/ios/orientation'
     require 'calabash/ios/server'
     require 'calabash/ios/text'
