+++ lib/calabash/ios/wait_for_condition.rb
